Added better handling when pid is not running

+++ admin/redbob_stopserver.sh
@@ -1,5 +1,9 @@
 #!/bin/bash
 
 pid=`pgrep -f "python prodserver.py --port=7001"`
-printf "killing process with pid=$pid\n"
-kill -9 $pid
+if [ "$pid" = "" ]; then
+	echo "No process found"
+else
+	echo "Killing process with pid=$pid\n"
+	kill -9 $pid
+fi
